Move to new file and add config to disable if needed
